app/requirements.txt
fastapi
uvicorn
azure-iot-hub
logging
os